Changed the path
The path of the pdf read was faulty as it only read the csv with the current month, so if the program would have been run first time in the month it would have given an error since there is no extracted file of the month. Change the path to select the latest file in from scraper cleaner.